Remove src folder, pytest-pythonpath requirement and instead rename src to model and make it a module
