+++ test-requirements.txt
@@ -1,4 +1,3 @@
 pytest==5.4.1
 pytest-cov==2.8.1
-pytest-pythonpath==0.7.3
 -r requirements.txt
